+++ schemas/feed_explained.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+
+
+class ExplainedFeed(BaseModel):
+    title: str
+    href: str | None
+    href_user: str
+    private: bool
+    frequency: str
+    notes: str
+    json: dict
